made privileges evaluator extensible through config
