+++ src/main/java/com/amazon/dlic/auth/http/kerberos/HTTPSpnegoAuthenticator.java
@@ -15,6 +15,7 @@
 
 package com.amazon.dlic.auth.http.kerberos;
 
+import java.io.IOException;
 import java.io.Serializable;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -33,6 +34,7 @@ import java.util.Set;
 import javax.security.auth.Subject;
 import javax.security.auth.login.LoginException;
 
+import com.amazon.opendistroforelasticsearch.security.support.ConfigConstants;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.ExceptionsHelper;
@@ -59,6 +61,7 @@ import com.amazon.dlic.auth.http.kerberos.util.KrbConstants;
 import com.amazon.opendistroforelasticsearch.security.auth.HTTPAuthenticator;
 import com.amazon.opendistroforelasticsearch.security.user.AuthCredentials;
 import com.google.common.base.Strings;
+import org.apache.hadoop.security.authentication.util.auth_to_name.KrbAuthToNameWrapper;
 
 public class HTTPSpnegoAuthenticator implements HTTPAuthenticator {
 
@@ -86,7 +89,7 @@ public class HTTPSpnegoAuthenticator implements HTTPAuthenticator {
             AccessController.doPrivileged(new PrivilegedAction<Void>() {
 
                 @Override
-                public Void run() {
+                public Void run() {log.debug("acceptor_keytab_filepath {}", acceptorKeyTabPath);
 
                     try {
                         if (settings.getAsBoolean("krb_debug", false)) {
@@ -156,6 +159,16 @@ public class HTTPSpnegoAuthenticator implements HTTPAuthenticator {
             log.debug("strip_realm_from_principal {}", stripRealmFromPrincipalName);
             log.debug("acceptor_principal {}", acceptorPrincipal);
             log.debug("acceptor_keytab_filepath {}", acceptorKeyTabPath);
+
+            if (Strings.isNullOrEmpty(settings.get(ConfigConstants.OPENDISTRO_SECURITY_KERBEROS_AUTH_TO_LOCAL_FILE_PATH))) {
+                log.warn("auth_to_local_file_path is empty, principal names will not be shortened");
+            } else {
+                try {
+                    KrbAuthToNameWrapper.init(settings.get(ConfigConstants.OPENDISTRO_SECURITY_KERBEROS_AUTH_TO_LOCAL_FILE_PATH), true);
+                } catch (IOException e) {
+                    log.warn("Error while initializing auth_to_local due to {}", e.getMessage(), e);
+                }
+            }
 
         } catch (Throwable e) {
             log.error("Cannot construct HTTPSpnegoAuthenticator due to {}", e.getMessage(), e);
@@ -255,8 +268,19 @@ public class HTTPSpnegoAuthenticator implements HTTPAuthenticator {
                     return new AuthCredentials("_incomplete_", (Object) outToken);
                 }
 
-
-                final String username = ((SimpleUserPrincipal) principal).getName();
+                String username = "";
+                try {
+                    String principalName = ((SimpleUserPrincipal) principal).getName();
+                    username = KrbAuthToNameWrapper.getName(principalName);
+                    log.debug ("username : " + username + " will be used for " + principalName);
+                    if (!Strings.isNullOrEmpty(principalName) && Strings.isNullOrEmpty(username)) {
+                        log.error("Wrong configuration in auth_to_local for the principal : {} . Please check. Using principal as username", principalName);
+                        username = principalName;
+                    }
+                } catch (IOException e) {
+                    log.warn("Exception while getting name from auth_to_local due to {}, will use principal as username : ", e.getMessage(), e);
+                    username = ((SimpleUserPrincipal) principal).getName();
+                }
 
                 if(username == null || username.length() == 0) {
                     log.error("Got empty or null user from kerberos. Normally this means that you acceptor principal {} does not match the server hostname", acceptorPrincipal);
